Rename warc_csv
Being consistent with naming of other report types
